+++ Runtime/Scripts/Tileset/ParseTileset.cs
@@ -3,11 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using SimpleJSON;
+using static Netherlands3D.Tiles3D.TileSetsConfigLoader.Config;
 
 namespace Netherlands3D.Tiles3D
 {
     public static class ParseTileset
     {
+        public static bool DebugLog = false;
+        
         private static readonly Dictionary<string, BoundingVolumeType> boundingVolumeTypes = new()
         {
             { "region", BoundingVolumeType.Region },
@@ -17,10 +20,10 @@ namespace Netherlands3D.Tiles3D
         public static readonly string[] SupportedExtensions = Array.Empty<string>(); //currently no extensions are supported
         
         internal static ReadSubtree subtreeReader;
-        internal static Tile ReadTileset(JSONNode rootnode)
+        internal static Tile ReadTileset(JSONNode rootnode, Read3DTileset tileset)
         {
             Tile root = new Tile();
-            
+            root.tileSet = tileset;
             TilingMethod tilingMethod = TilingMethod.ExplicitTiling;
             double[] transformValues = new double[16] { 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0 };
             JSONNode transformNode = rootnode["transform"];
@@ -43,8 +46,10 @@ namespace Netherlands3D.Tiles3D
             switch (tilingMethod)
             {
                 case TilingMethod.ExplicitTiling:
-                    Debug.Log("Explicit tiling");
+                    if(DebugLog)
+                        Debug.Log("Explicit tiling");
                     Tile rootTile = new Tile();
+                    rootTile.tileSet = tileset;
                     rootTile.tileTransform = root.tileTransform;
                     root = ReadExplicitNode(rootnode, rootTile);
                     
@@ -53,6 +58,7 @@ namespace Netherlands3D.Tiles3D
                     if (rootTile.children.Count==0 )
                     {
                         Tile childTile = new Tile();
+                        childTile.tileSet = tileset;
                         childTile = ReadExplicitNode(rootnode, childTile);
                         childTile.parent = rootTile;
                         rootTile.geometricError += 10;
@@ -62,16 +68,18 @@ namespace Netherlands3D.Tiles3D
                     }
                     break;
                 case TilingMethod.ImplicitTiling:
-                    Debug.Log("Implicit tiling"); 
+                    if(DebugLog)
+                        Debug.Log("Implicit tiling"); 
                     rootTile = new Tile();
+                    rootTile.tileSet = tileset;
                     rootTile = ReadExplicitNode(rootnode, rootTile);
                     rootTile.level = 0;
                     rootTile.X = 0;
                     rootTile.Y = 0;
                     rootTile.transform = root.transform;
-                    
-                    ReadImplicitTiling(rootnode,rootTile);
                     rootTile.tileTransform = root.tileTransform;
+                    ReadImplicitTiling(rootnode,rootTile);
+
                     rootTile.transform = root.transform;
                     root = rootTile;
                     break;
@@ -129,6 +137,7 @@ namespace Netherlands3D.Tiles3D
                 for (int i = 0; i < childrenNode.Count; i++)
                 {
                     var childTile = new Tile();
+                    childTile.tileSet = tile.tileSet;
                     childTile.transform = tile.transform;
                     childTile.parent = tile;
                     tile.children.Add(ReadExplicitNode(childrenNode[i], childTile));
@@ -225,8 +234,8 @@ namespace Netherlands3D.Tiles3D
             implicitTilingSettings.subtreeUri = implicitTilingNode["subtrees"]["uri"].Value;
             subtreeReader.settings = implicitTilingSettings;
             
-
-            Debug.Log("Load subtree: " + "");
+            if(DebugLog)
+                Debug.Log("Load subtree: " + "");
             if (parentTile.level==0)
             {
                 parentTile.contentUri = implicitTilingSettings.subtreeUri.Replace("{level}", parentTile.level.ToString()).Replace("{x}", parentTile.X.ToString()).Replace("{y}", parentTile.Y.ToString());