+++ CHANGELOG.md
@@ -5,8 +5,19 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.10.8] - 05-08-2025
-- Added support for screen space error calculation for orthographic camera
+## [2.2.0] - 31-07-2025
+- Added events for when tile content is loaded
+
+## [2.1.0] - 10-07-2025
+- Added function SetCoordinateSystem to Read3DTiles.cs
+
+## [2.0.0] - 16-06-2025
+- This package now uses Unity's default GLTFast package 
+- Added NL3D material generator to work with custom shaders (mostly catered to NL3D)
+- Made debug logs optional
+
+## [1.10.8] - 22-05-2025
+- properly clearing the used texture by looking up the right texture property in the shader
 
 ## [1.10.7] - 23-04-2025
 - Added console logging to ImportGlb to improve troubleshooting during the GLB import process.